Add delete confirmation

+++ src/components/Detail/AccountDetail.js
@@ -7,9 +7,11 @@ export default function AccountDetail() {
     // slug -> accountid
     const { slug } = useParams();
     const [playerData, setPlayerData] = useState([]);
+    const [playerGameAcc, setPlayerGameAcc] = useState([]);
 
     useEffect(() => {
         fetchPlayer();
+        fetchInGameAcc();
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [])
 
@@ -23,20 +25,32 @@ export default function AccountDetail() {
             .catch(err => alert(err));
     }
 
+    const fetchInGameAcc = () => {
+        // const sql = `select * from game_server where gameid=${slug}`
+        axios.get(`https://gamedb-api-service.up.railway.app/api/get-playerlist?accountid=${slug}`)
+            // if fetching game server successfully
+            .then(response => {
+                setPlayerGameAcc(response.data);
+            })
+            .catch(err => alert(err));
+    }
+
     const deleteAccount = (e) => {
         e.preventDefault();
-        const sql = `delete from account where accountid=${slug}`;
-        axios.post(`https://gamedb-api-service.up.railway.app/api/execute-query`, { sql })
-            // if delete successfully
-            .then(response => {
-                if (response.data['error'])
-                {
-                    alert(response.data.error); return;
-                }
-                alert("This account has been deleted successfully")
-                window.location = "/accounts";
-            })
-            .catch(error => alert(error));
+        if (window.confirm("Are you sure you want to delete this game?")) {
+            const sql = `delete from account where accountid=${slug}`;
+            axios.post(`https://gamedb-api-service.up.railway.app/api/execute-query`, { sql })
+                // if delete successfully
+                .then(response => {
+                    if (response.data['error'])
+                    {
+                        alert(response.data.error); return;
+                    }
+                    alert("This account has been deleted successfully")
+                    window.location = "/accounts";
+                })
+                .catch(error => alert(error));
+        }
     }
 
     return (
@@ -53,6 +67,26 @@ export default function AccountDetail() {
                             <p>Email : {item.email}</p>
                             <p>Password : {item.password}</p>
                             <p>Account Register Date : {item.accountregisterdate.substring(0, item.accountregisterdate.indexOf("T"))}</p>
+                        </div>
+                    ))
+                ) : (
+                    // Render message if no data available
+                    <p>Data Fetching...</p>
+                )}
+            </div>
+
+            <div className="game-item">
+                {playerGameAcc.length > 0 ? (
+                    // Render if server available
+                    playerGameAcc.map(item => (
+                        <div key={item.accountid}>
+                            <p className='bold'>Game Account ID: {item.gameaccountid}</p>
+                            <p>Game ID: {item.gameid}</p>
+                            <p>Game Server ID: {item.gameserverid}</p>
+                            <p>In-game name: {item.ingamename}</p>
+                            <p>Account level: {item.accountlevel}</p>
+                            <p>Register Date: {item.ingameregisterdate}</p>
+                            <br/>
                         </div>
                     ))
                 ) : (
