+++ src/components/Detail/PlayerDetail.js
@@ -5,11 +5,23 @@ import axios from 'axios';
 export default function PlayerDetail() {
     const { gameid, slug } = useParams();
     const [playerData, setPlayerData] = useState([]);
+    // const [mainAcc, setMainAcc] = useState([]);
 
     useEffect(() => {
         fetchPlayer();
+        // fetchMainAcc();
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [])
+
+    // const fetchMainAcc = () => {
+    //     // const sql = `select * from game_server where gameid=${slug}`
+    //     axios.get(`https://gamedb-api-service.up.railway.app/api/get-accountdetails?accountid=${playerData.length>0? playerData[0].accountid: ""}`)
+    //         // if fetching game server successfully
+    //         .then(response => {
+    //             setMainAcc(response.data);
+    //         })
+    //         .catch(err => alert(err));
+    // }
 
     const fetchPlayer = () => {
         // const sql = `select * from game_server where gameid=${slug}`
@@ -23,23 +35,43 @@ export default function PlayerDetail() {
 
     const deletePlayer = (e) => {
         e.preventDefault();
-        const sql = `delete from ingame_account where gameaccountid=${slug}`;
-        axios.post(`https://gamedb-api-service.up.railway.app/api/execute-query`, { sql })
-            // if delete successfully
-            .then(response => {
-                if (response.data['error'])
-                {
-                    alert(response.data.error); return;
-                }
-                alert("This player has been deleted successfully")
-                window.location = "/players";
-            })
-            .catch(error => alert(error));
+        if (window.confirm("Are you sure you want to delete this game?")) {
+            const sql = `delete from ingame_account where gameaccountid=${slug}`;
+            axios.post(`https://gamedb-api-service.up.railway.app/api/execute-query`, { sql })
+                // if delete successfully
+                .then(response => {
+                    if (response.data['error'])
+                    {
+                        alert(response.data.error); return;
+                    }
+                    alert("This player has been deleted successfully")
+                    window.location = "/players";
+                })
+                .catch(error => alert(error));
+        }
     }
 
     return (
         <main>
-            <h1>Player Detail: Game Account ID = {slug}</h1>
+            <h1>{playerData.length>0? playerData[0].ingamename: ""}</h1>
+
+            {/* <div className="game-item">
+                {mainAcc.length > 0 ? (
+                    // Render if server available
+                    mainAcc.map(item => (
+                        <div key={item.accountid}>
+                            <p className='bold'>Account ID: {item.accountid}</p>
+                            <p>Username : {item.username}</p>
+                            <p>Email : {item.email}</p>
+                            <p>Password : {item.password}</p>
+                            <p>Account Register Date : {item.accountregisterdate.substring(0, item.accountregisterdate.indexOf("T"))}</p>
+                        </div>
+                    ))
+                ) : (
+                    // Render message if no data available
+                    <p>Data Fetching...</p>
+                )}
+            </div> */}
 
             <div className="game-item">
                 {playerData.length > 0 ? (
