+++ src/components/Detail/ServLocDatial.js
@@ -19,25 +19,26 @@ export default function ServLocDatial() {
             .catch(err => alert(err));
     }
 
-     const deleteServerLocation = (e) => {
-         e.preventDefault();
-         const sql = `delete from server_location where serverlocationid=${slug}`;
-         axios.post(`https://gamedb-api-service.up.railway.app/api/execute-query`, { sql })
-             // if delete successfully
-             .then(response => {
-                if (response.data['error'])
-                {
-                    alert(response.data.error); return;
-                }
-                 alert("This server location has been deleted successfully")
-                 window.location = "/server-locations";
-             })
-             .catch(error => alert(error));
-     }
+    const deleteServerLocation = (e) => {
+        e.preventDefault();
+        if (window.confirm("Are you sure you want to delete this game?")) {
+            const sql = `delete from server_location where serverlocationid=${slug}`;
+            axios.post(`https://gamedb-api-service.up.railway.app/api/execute-query`, { sql })
+                // if delete successfully
+                .then(response => {
+                    if (response.data['error']) {
+                        alert(response.data.error); return;
+                    }
+                    alert("This server location has been deleted successfully")
+                    window.location = "/server-locations";
+                })
+                .catch(error => alert(error));
+        }
+    }
 
     return (
         <main>
-            <h1>Server Location {serverloc.length>0? serverloc[0].region: ''}</h1>
+            <h1>Server Location {serverloc.length > 0 ? serverloc[0].region : ''}</h1>
 
             <div>
                 {serverloc.length > 0 ? (
@@ -47,7 +48,7 @@ export default function ServLocDatial() {
                             <p>Region: {item.region}</p>
                             <p>Colocation country: {item.colocation_country}</p>
                             <p>Colocation company: {item.colocation_company}</p>
-                            
+
                         </div>
                     ))
                 ) : (
@@ -58,15 +59,15 @@ export default function ServLocDatial() {
             <Link to={`/edit-serverlocation/${slug}`}>
                 <button className='edit-btn'>Edit location</button>
             </Link>
-            <br/>
+            <br />
             <Link to={"/server-locations"}>
                 <button className='delete-btn' onClick={deleteServerLocation}>Delete location</button>
             </Link>
-            <br/>
+            <br />
             <Link to={"/server-locations"}>
                 <button className='nice_butt_on'>All server locations</button>
             </Link>
-            <br/>
+            <br />
             <Link to={"/"}>
                 <button className='nice_butt_on'>Home</button>
             </Link>
